requirements.txt
fastapi
uvicorn[standard]
flask
celery
redis
python-dotenv
rich
typer
pandas
numpy
libsass
matplotlib
sqlalchemy[asyncio]
asyncpg
alembic
hydra-core
jinja2
quarto-cli
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Babel
Authlib
markdown        # para convertir Markdown a HTML
requests        # para proxy de métricas
fastapi
uvicorn
python-dotenv
authlib
starlette
python-multipart
jinja2   # si necesitas templates